Auto-Coder: Fix issues based on GitHub Actions logs (PR #3023)
