+++ client/e2e/yjs/20-yjs-order-sync-7f2c1d3e.spec.ts
@@ -60,7 +60,7 @@ test("order: p1 connect->set then p2 connect for initial sync", async ({ browser
     // Wait for both provider.synced and actual data to be available using the test utility function
     const value = await p2.evaluate(async () => {
         // @ts-expect-error - Browser context import resolved by Vite
-        const { waitForSyncedAndDataForTest } = await import("/src/lib/yjs/testHelpers.ts");
+        const { waitForSyncedAndDataForTest } = await import("/src/lib/yjs/browserTestHelpers.ts");
 
         const provider = (window as any).__PROVIDER2__;
